Use tabs for formatting only based on expandtab setting
The 'smarttab' option does not affect whether tabs or spaces are used
for indentation, and thus should not be used to determine whether the
`-tabs` flag should be used.

The 'smarttab' option only determines whether a <Tab> at the in front of
a line should align to the indentation level. From the documentation for
'smarttab':

	What gets inserted (a <Tab> or spaces) depends on the 'expandtab'
	option.